Fresh restart!
Restarting the code all over. The purpose is to change some fundamental
concepts about the game and how things are done. This will eventually
lead to the implementation of many features not available on the
original code base.

It also seemed as a good chance to try out a new framework: SFML,
which appears to be better organized, more OO and also contains some
useful modules for dealing with things such as TrueType Font rendering
and audio playing.